finished skyband query!

+++ sky_point.h
@@ -17,7 +17,7 @@
 
 typedef struct SkyPoint {
     int dim;                    /* Point dimension */
-    int32 **data;               /* Stored data of this point */
+    double **data;               /* Stored data of this point */
     char *bitmap;               /* Bitmap of this point */
     int cnt_domi;               /* The dominated count of this point */
     int index;                  /* The index of this point */
